Test fixed

+++ src/main/java/com/ambit/occupancy/UsagePair.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class UsagePair {
-    private int premiumSum;
-    private int economySum;
+    private long premiumSum;
+    private long economySum;
 
-    public UsagePair(int premiumSum, int economySum) {
+    public UsagePair(long premiumSum, long economySum) {
 
         this.premiumSum = premiumSum;
         this.economySum = economySum;
